Copied SwitchToAscending to main,cpp
